Authentication/src/main/java/nl/tudelft/sem/template/authentication/SignUp.java
package nl.tudelft.sem.template.authentication;


public class SignUp {

    private transient String netId;
    private transient String firstName;
    private transient  String lastName;
    private transient String password;
    private transient String role;

    public SignUp() {
        super();
    }



    public String getNetId() {
        return netId;
    }

    public void setNetId(String netId) {
        this.netId = netId;
    }

    public String getFirstName() {
        return firstName;
    }

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public String getLastName() {
        return lastName;
    }

    public void setLastName(String lastName) {
        this.lastName = lastName;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }

    public String getRole() {
        return this.role;
    }

    public void setRole(String role) {
        this.role = role;
    }
}

